adding redis_cms

+++ src/types/cms.h
@@ -21,38 +21,43 @@
 #pragma once
 
 #include <assert.h>
-#include <atomic>
-#include <vector>
-#include "vendor/murmurhash2.h"
-#include <iostream>
 
+#include <atomic>
+#include <iostream>
+#include <vector>
+
+#include "vendor/murmurhash2.h"
 
 class CMSketch {
-  public:
-    explicit CMSketch(uint32_t width = 0, uint32_t depth = 0, uint64_t counter = 0, std::vector<uint32_t> array = {}) 
-        : width_(width), depth_(depth), counter_(counter), array_(array.empty() ? std::vector<uint32_t>(width * depth, 0) : std::move(array)) {}
+ public:
+  explicit CMSketch(uint32_t width = 0, uint32_t depth = 0, uint64_t counter = 0, std::vector<uint32_t> array = {})
+      : width_(width),
+        depth_(depth),
+        counter_(counter),
+        array_(array.empty() ? std::vector<uint32_t>(width * depth, 0) : std::move(array)) {}
 
-    ~CMSketch() = default;
+  ~CMSketch() = default;
 
-    static CMSketch* NewCMSketch(size_t width, size_t depth) { return new CMSketch(width, depth); }
+  static CMSketch* NewCMSketch(size_t width, size_t depth) { return new CMSketch(width, depth); }
 
-    static void CMSDimFromProb(double error, double delta, size_t &width, size_t &depth);
+  static void CMSDimFromProb(double error, double delta, uint32_t& width, uint32_t& depth);
 
-    size_t IncrBy(const char* item, size_t item_len, size_t value);
+  size_t IncrBy(const char* item, size_t item_len, size_t value);
 
-    size_t Query(const char* item, size_t item_len) const;
+  size_t Query(const char* item, size_t item_len) const;
 
-    static int Merge(CMSketch* dest, size_t quantity, const std::vector<const CMSketch*>& src, const std::vector<long long>& weights);
+  static int Merge(CMSketch* dest, size_t quantity, const std::vector<const CMSketch*>& src,
+                   const std::vector<long long>& weights);
 
-    struct MergeParams {
-      CMSketch* dest;
-      size_t num_keys;
-      std::vector<const CMSketch*> cms_array;
-      std::vector<long long> weights;
+  struct MergeParams {
+    CMSketch* dest;
+    size_t num_keys;
+    std::vector<const CMSketch*> cms_array;
+    std::vector<long long> weights;
   };
 
   int CMSMergeParams(const MergeParams& params);
-  
+
   uint64_t& GetCounter() { return counter_; }
   std::vector<uint32_t>& GetArray() { return array_; }
 
@@ -62,13 +67,16 @@ class CMSketch {
   size_t GetWidth() const { return width_; }
   size_t GetDepth() const { return depth_; }
 
-  private:
-    size_t width_;
-    size_t depth_;
-    uint64_t counter_;
-    std::vector<uint32_t> array_;
+ private:
+  size_t width_;
+  size_t depth_;
+  uint64_t counter_;
+  std::vector<uint32_t> array_;
 
-    static uint32_t hllMurMurHash64A(const char* item, size_t item_len, size_t i) { return HllMurMurHash64A(item, static_cast<int>(item_len), i); }
+  static uint32_t hllMurMurHash64A(const char* item, size_t item_len, size_t i) {
+    return HllMurMurHash64A(item, static_cast<int>(item_len), i);
+  }
 
-    static int checkOverflow(CMSketch* dest, size_t quantity, const std::vector<const CMSketch*>& src, const std::vector<long long>& weights);
+  static int checkOverflow(CMSketch* dest, size_t quantity, const std::vector<const CMSketch*>& src,
+                           const std::vector<long long>& weights);
 };
